Convert KubeServer and KubernetesCluster to handler format (#63084)
* remove old kube handlers

* remove kube cluster and server collection

* add new kube server handler

* add new kube cluster handler

* use constructors

* update tests

* fix fmt

* update descriptions
