+++ tool/tctl/common/kube_command.go
@@ -36,6 +36,7 @@ import (
 	"github.com/gravitational/teleport/lib/utils"
 	commonclient "github.com/gravitational/teleport/tool/tctl/common/client"
 	tctlcfg "github.com/gravitational/teleport/tool/tctl/common/config"
+	"github.com/gravitational/teleport/tool/tctl/common/resources"
 )
 
 // KubeCommand implements "tctl kube" group of commands.
@@ -108,14 +109,14 @@ func (c *KubeCommand) ListKube(ctx context.Context, clt *authclient.Client) erro
 		return trace.Wrap(err)
 	}
 
-	coll := &kubeServerCollection{servers: kubes}
+	coll := resources.NewKubeServerCollection(kubes)
 	switch c.format {
 	case teleport.Text:
 		return trace.Wrap(coll.WriteText(os.Stdout, c.verbose))
 	case teleport.JSON:
-		return trace.Wrap(coll.WriteJSON(os.Stdout))
+		return trace.Wrap(writeJSON(coll, os.Stdout))
 	case teleport.YAML:
-		return trace.Wrap(coll.WriteYAML(os.Stdout))
+		return trace.Wrap(writeYAML(coll, os.Stdout))
 	default:
 		return trace.BadParameter("unknown format %q", c.format)
 	}
